Crate logos_codegen

Logos

This is a #[derive] macro crate, for documentation go to main crate.

Functions

generate

Generate a Logos implementation for the given struct, provided as a stream of rust tokens.

strip_attributes

Strip all logos attributes from the given struct, allowing it to be used in code without logos-derive present.