logos_codegen::generator
Struct Generator

pub struct Generator<'a> {
    name: &'a Ident,
    this: &'a TokenStream,
    root: NodeId,
    graph: &'a Graph<Leaf<'a>>,
    meta: Meta,
    rendered: TokenStream,
    fns: FnvHashSet<(NodeId, Context)>,
    idents: FnvHashMap<(NodeId, Context), Ident>,
    gotos: FnvHashMap<(NodeId, Context), TokenStream>,
    tests: FnvHashMap<Vec<Range>, Ident>,
    tables: TableStack,
}

Fields
name: &'a Ident

Name of the type we are implementing the Logos trait for

this: &'a TokenStream

Name of the type with any generics it might need

root: NodeId

Id to the root node

graph: &'a Graph<Leaf<'a>>

Reference to the graph with all of the nodes

meta: Meta

Meta data collected for the nodes

rendered: TokenStream

Buffer with functions growing during generation

fns: FnvHashSet<(NodeId, Context)>

Set of functions that have already been rendered

idents: FnvHashMap<(NodeId, Context), Ident>

Function name identifiers

gotos: FnvHashMap<(NodeId, Context), TokenStream>

Local function calls. Note: a call might change its context, so we can’t use idents for this purpose.

tests: FnvHashMap<Vec<Range>, Ident>

Identifiers for helper functions matching a byte to a given set of ranges

tables: TableStack

Related to above, table stack manages tables that need to be

Implementations

impl Generator<'_>

pub fn generate_fork( &mut self, this: NodeId, fork: &Fork, ctx: Context, ) -> TokenStream

fn generate_fork_jump_table( &mut self, this: NodeId, fork: &Fork, targets: FnvHashMap<NodeId, Vec<Range>>, ctx: Context, ) -> TokenStream

fn fork_end(&self, this: NodeId, miss: &TokenStream) -> TokenStream

fn fork_read( &self, this: NodeId, end: TokenStream, ctx: &mut Context, ) -> (TokenStream, TokenStream)

fn generate_fast_loop(&mut self, fork: &Fork, ctx: Context) -> TokenStream

pub fn fast_loop_macro() -> TokenStream

impl Generator<'_>

pub fn generate_leaf(&mut self, leaf: &Leaf<'_>, ctx: Context) -> TokenStream

impl Generator<'_>

pub fn generate_rope(&mut self, rope: &Rope, ctx: Context) -> TokenStream

impl<'a> Generator<'a>

pub fn new( name: &'a Ident, this: &'a TokenStream, root: NodeId, graph: &'a Graph<Leaf<'_>>, ) -> Self

pub fn generate(self) -> TokenStream

fn generate_fn(&mut self, id: NodeId, ctx: Context)

fn goto(&mut self, id: NodeId, ctx: Context) -> &TokenStream

fn generate_ident(&mut self, id: NodeId, ctx: Context) -> &Ident

fn generate_test(&mut self, ranges: Vec<Range>) -> &Ident

Returns an identifier to a function that matches a byte to any of the provided ranges. This will generate either a simple match expression, or use a lookup table internally.