logos_codegen
Module graph

Re-exports

pub use self::fork::Fork;

pub use self::meta::Meta;

pub use self::range::Range;

pub use self::rope::Rope;

Modules

fork 🔒

impls 🔒

meta 🔒

range 🔒

regex 🔒

rope 🔒

Structs

DeferredMerge

When attempting to merge two nodes, one of which was not yet created, we can record such attempt, and execute the merge later on when the awaiting has been inserted into the graph.

DisambiguationError

Disambiguation error during the attempt to merge two leaf nodes with the same priority

Graph

Merge 🔒

Merge key used to lookup whether two nodes have been previously mered, so we can avoid duplicating merges, potentially running into loops that blow out the stack.

NodeId

Id of a Node in the graph. NodeId can be referencing an empty slot that is going to be populated later in time.

ReservedId

Unique reserved NodeId that is guaranteed to point to an empty allocated slot in the graph. It’s safe to create multiple NodeId copies of ReservedId, however API users should never be able to clone a ReservedId, or create a new one from NodeId.

Enums

Node

Traits

Disambiguate

Trait to be implemented on Leaf nodes in order to disambiguate between them.