logos_codegen::graph
Enum Node

pub enum Node<Leaf> {
    Fork(Fork),
    Rope(Rope),
    Leaf(Leaf),
}

Variants

Fork(Fork)

Fork node, can lead to more than one state

Rope(Rope)

Rope node, can lead to one state on match, one state on miss

Leaf(Leaf)

Leaf node, terminal state

Implementations

impl<Leaf> Node<Leaf>

pub fn miss(&self) -> Option<NodeId>

fn eq(&self, other: &Node<Leaf>) -> bool

fn shake(&self, graph: &Graph<Leaf>, filter: &mut [bool])

pub fn unwrap_leaf(&self) -> &Leaf