logos_codegen::graph::regex
Function is_ascii

fn is_ascii(class: &ClassUnicode, repeated: bool) -> bool

Return whether current class unicode is ascii.

Because unicode ranges are iterated in increasing order, it is only necessary to check the last range.

If the check is performed in a repetition, a fast path is used by checking if end of range is 0x0010_FFFF.