logos_codegen::graph
Struct Graph

pub struct Graph<Leaf> {
    nodes: Vec<Option<Node<Leaf>>>,
    merges: BTreeMap<Merge, NodeId>,
    hashes: BTreeMap<u64, NodeId>,
    errors: Vec<DisambiguationError>,
    deferred: Vec<DeferredMerge>,
}

Fields
nodes: Vec<Option<Node<Leaf>>>

Internal storage of all allocated nodes. Once a node is put here, it should never be mutated.

merges: BTreeMap<Merge, NodeId>

When merging two nodes into a new node, we store the two entry keys and the result, so that we don’t merge the same two nodes multiple times.

Most of the time the entry we want to find will be the last one that has been inserted, so we can use a vec with reverse order search to get O(1) searches much faster than any *Map.

hashes: BTreeMap<u64, NodeId>

Another map used for accounting. Before .pushing a new node onto the graph (inserts are exempt), we hash it and find if an identical(!) node has been created before.

errors: Vec<DisambiguationError>

Instead of handling errors over return types, opt to collect them internally.

deferred: Vec<DeferredMerge>

Deferred merges. When when attempting to merge a node with an empty reserved slot, the merging operation is deferred until the reserved slot is populated. This is a stack that keeps track of all such deferred merges

Implementations

impl<Leaf: Disambiguate + Debug> Graph<Leaf>

pub fn regex(&mut self, mir: Mir, then: NodeId) -> NodeId

fn parse_mir( &mut self, mir: &Mir, then: NodeId, miss: Option<NodeId>, reserved: Option<ReservedId>, repeated: bool, ) -> NodeId

fn insert_or_push<N>(&mut self, id: Option<ReservedId>, node: N) -> NodeId
where N: Into<Node<Leaf>>,

impl<Leaf> Graph<Leaf>

pub fn new() -> Self

pub fn errors(&self) -> &[DisambiguationError]

fn next_id(&self) -> NodeId

pub fn reserve(&mut self) -> ReservedId

Reserve an empty slot for a node on the graph and return an id for it. ReservedId cannot be cloned, and must be consumed by calling insert on the graph.

pub fn insert<N>(&mut self, reserved: ReservedId, node: N) -> NodeId
where N: Into<Node<Leaf>>, Leaf: Disambiguate,

Insert a node at a given, previously reserved id. Returns the inserted NodeId.

pub fn push<B>(&mut self, node: B) -> NodeId
where B: Into<Node<Leaf>>,

Push a node onto the graph and get an id to it. If an identical node has already been pushed on the graph, it will return the id of that node instead.

fn push_unchecked(&mut self, node: Node<Leaf>) -> NodeId

fn find_merged(&self, a: NodeId, b: NodeId) -> Option<NodeId>

If nodes a and b have been already merged, return the NodeId of the node they have been merged into.

fn set_merged(&mut self, a: NodeId, b: NodeId, product: NodeId)

Mark that nodes a and b have been merged into product.

This will also mark merging a and product, as well as b and product into product, since those are symmetric operations.

This is necessary to break out asymmetric merge loops.

pub fn merge(&mut self, a: NodeId, b: NodeId) -> NodeId
where Leaf: Disambiguate,

Merge the nodes at id a and b, returning a new id.

fn merge_unchecked( &mut self, a: NodeId, b: NodeId, reserved: ReservedId, ) -> NodeId
where Leaf: Disambiguate,

Unchecked merge of a and b. This fn assumes that a and b are not pointing to empty slots.

fn merge_rope(&mut self, rope: Rope, other: NodeId) -> Option<Rope>
where Leaf: Disambiguate,

pub fn fork_off(&mut self, id: NodeId) -> Fork
where Leaf: Disambiguate,

pub fn nodes(&self) -> &[Option<Node<Leaf>>]

pub fn shake(&mut self, root: NodeId)

Find all nodes that have no references and remove them.

pub fn get(&self, id: NodeId) -> Option<&Node<Leaf>>