logos_codegen
Module mir

Structs

ClassUnicode

A set of characters represented by Unicode scalar values.

DOT_BYTES 🔒

DOT_UTF8 🔒

Literal

The high-level intermediate representation of a literal.

Enums

Class

The high-level intermediate representation of a character class.

Mir

Middle Intermediate Representation of the regex, built from regex_syntax’s Hir. The goal here is to strip and canonicalize the tree, so that we don’t have to do transformations later on the graph, with the potential of running into looping references.