logos_codegen::mir
Enum Mir

pub enum Mir {
    Empty,
    Loop(Box<Mir>),
    Maybe(Box<Mir>),
    Concat(Vec<Mir>),
    Alternation(Vec<Mir>),
    Class(Class),
    Literal(Literal),
}

Middle Intermediate Representation of the regex, built from regex_syntaxโs Hir. The goal here is to strip and canonicalize the tree, so that we donโt have to do transformations later on the graph, with the potential of running into looping references.

Variants

Empty

Loop(Box<Mir>)

Maybe(Box<Mir>)

Concat(Vec<Mir>)

Alternation(Vec<Mir>)

Class(Class)

Literal(Literal)

Implementations

impl Mir

pub fn utf8(source: &str) -> Result<Mir, Error>

pub fn utf8_ignore_case(source: &str) -> Result<Mir, Error>

pub fn binary(source: &str) -> Result<Mir, Error>

pub fn binary_ignore_case(source: &str) -> Result<Mir, Error>

pub fn priority(&self) -> usize