logos_codegen::parser
Struct Parser

pub struct Parser {
    pub errors: Errors,
    pub mode: Mode,
    pub source: Option<TokenStream>,
    pub skips: Vec<Literal>,
    pub extras: MaybeVoid,
    pub error_type: MaybeVoid,
    pub subpatterns: Subpatterns,
    pub logos_path: Option<TokenStream>,
    types: TypeParams,
}

Fields
errors: Errorsmode: Modesource: Option<TokenStream>skips: Vec<Literal>extras: MaybeVoiderror_type: MaybeVoidsubpatterns: Subpatternslogos_path: Option<TokenStream>types: TypeParams
Implementations

impl Parser

pub fn parse_generic(&mut self, param: GenericParam)

pub fn generics(&mut self) -> Option<TokenStream>

fn parse_attr(&mut self, attr: &mut Attribute) -> Option<AttributeParser>

pub fn try_parse_logos(&mut self, attr: &mut Attribute)

Try to parse the main #[logos(...)], does nothing if the attribute’s name isn’t logos.

pub fn parse_literal(&mut self, lit: Lit) -> Option<Literal>

pub fn parse_definition(&mut self, attr: &mut Attribute) -> Option<Definition>

Parse attribute definition of a token:

#[token(literal[, callback])]

#[regex(literal[, callback])]

fn parse_callback(&mut self, tokens: TokenStream) -> Option<Callback>

pub fn get_type(&self, ty: &mut Type) -> TokenStream

Checks if ty is a declared generic param, if so replaces it with a concrete type defined using #[logos(type T = Type)]

If no matching generic param is found, all lifetimes are fixed to the source lifetime

pub fn err<M>(&mut self, message: M, span: Span) -> &mut Errors
where M: Into<Cow<'static, str>>,