Crate logos_codegen

Logos

This is a #[derive] macro crate, for documentation go to main crate.

Modules

error ğŸ”’

generator ğŸ”’

graph ğŸ”’

leaf ğŸ”’

macros ğŸ”’

mir ğŸ”’

parser ğŸ”’

util ğŸ”’

Constants

ERROR_ATTR ğŸ”’

LOGOS_ATTR ğŸ”’

REGEX_ATTR ğŸ”’

TOKEN_ATTR ğŸ”’

Functions

generate

Generate a Logos implementation for the given struct, provided as a stream of rust tokens.

is_logos_attr ğŸ”’

strip_attributes

Strip all logos attributes from the given struct, allowing it to be used in code without logos-derive present.

strip_attrs_from_vec ğŸ”’

strip_wrapping_parens ğŸ”’